requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2

# API and data fetching
requests==2.31.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Visualization
plotly==5.18.0
streamlit==1.29.0

# Time series modeling
prophet==1.1.5
statsmodels==0.14.1
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# MLOps (Part 2 - commented out for now)
# mlflow==2.9.2
# evidently==0.4.11
# prometheus-client==0.19.0
